abacus/scripts/requirements.txt
marker-pdf>=1.7.0
click>=8.1.0
anthropic>=0.40.0
python-dotenv>=1.0.0
